Added LicenseeService feature and Fixed Validation for Floating Licenses
I added the ability to pass additional attributes to the licensee
validation request. This means it is now possible to validate and
check-in and check-out floating licenses which was not previously
possible.

I also added a LicenseeService module which was based on the existing
ProductService. All licensee REST requests can now be made via the API.

However I did spot a problem with the Mapper class, and provided a note
in licensee.h about it.